+++ src/PlayerTrackerKernel.java
@@ -1,3 +1,5 @@
+import java.util.Set;
+
 import components.standard.Standard;
 
 /**
@@ -11,4 +13,6 @@ public interface PlayerTrackerKernel extends Standard<PlayerTrackerKernel> {
     void updateStat(String playerName, String stat, int value);
 
     int getStat(String playerName, String stat);
+
+    Set<String> allPlayers();
 }